:recycle: refactor: MainLayout component created
